Rename subscription types
