+++ subscription/index.ts
@@ -1,5 +1,10 @@
 import type { Key, SWRHook, Middleware, SWRConfiguration, SWRConfig } from 'swr'
-
+import type {
+  SWRSubscriptionOptions,
+  SWRSubscription,
+  SWRSubscriptionResponse,
+  SWRSubscriptionHook
+} from './types'
 import useSWR from 'swr'
 import {
   withMiddleware,
@@ -7,37 +12,6 @@ import {
   useIsomorphicLayoutEffect,
   createCacheHelper
 } from 'swr/_internal'
-
-export type SWRSubNext<Data = any, Error = any> = {
-  next: (err?: Error | null, data?: Data) => void
-}
-
-export type SWRSubscription<
-  SWRSubKey extends Key = Key,
-  Data = any,
-  Error = any
-> = SWRSubKey extends () => infer Arg | null | undefined | false
-  ? (key: Arg, { next }: SWRSubNext<Data, Error>) => void
-  : SWRSubKey extends null | undefined | false
-  ? never
-  : SWRSubKey extends infer Arg
-  ? (key: Arg, { next }: SWRSubNext<Data, Error>) => void
-  : never
-
-export type SWRSubscriptionResponse<Data = any, Error = any> = {
-  data?: Data
-  error?: Error
-}
-
-export type SWRSubscriptionHook = <
-  Data = any,
-  Error = any,
-  SWRSubKey extends Key = Key
->(
-  key: SWRSubKey,
-  subscribe: SWRSubscription<SWRSubKey, Data, Error>,
-  config?: SWRConfiguration
-) => SWRSubscriptionResponse<Data, Error>
 
 // [subscription count, disposer]
 type SubscriptionStates = [Map<string, number>, Map<string, () => void>]
@@ -149,3 +123,10 @@ const useSWRSubscription = withMiddleware(
 ) as SWRSubscriptionHook
 
 export default useSWRSubscription
+
+export type {
+  SWRSubscription,
+  SWRSubscriptionOptions,
+  SWRSubscriptionResponse,
+  SWRSubscriptionHook
+}
