+++ test/type/sub.ts
@@ -1,43 +1,46 @@
 import useSWRSubscription from 'swr/subscription'
-import type { SWRSubNext, SWRSubscription } from 'swr/subscription'
+import type { SWRSubscriptionOptions, SWRSubscription } from 'swr/subscription'
 import { expectType, truthy } from './utils'
 
 export function useTestSubscription() {
-  useSWRSubscription('key', (key, { next: _ }: SWRSubNext<string, Error>) => {
-    expectType<'key'>(key)
-    return () => {}
-  })
+  useSWRSubscription(
+    'key',
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
+      expectType<'key'>(key)
+      return () => {}
+    }
+  )
   useSWRSubscription(
     truthy() ? 'key' : undefined,
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<'key'>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     ['key', 1],
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<[string, number]>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     truthy() ? ['key', 1] : undefined,
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<[string, number]>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     { foo: 'bar' },
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<{ foo: string }>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     truthy() ? { foo: 'bar' } : undefined,
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<{ foo: string }>(key)
       return () => {}
     }
@@ -45,42 +48,42 @@ export function useTestSubscription() {
 
   useSWRSubscription(
     () => 'key',
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<string>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     () => (truthy() ? 'key' : undefined),
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<'key'>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     () => ['key', 1],
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<[string, number]>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     () => (truthy() ? ['key', 1] : undefined),
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<[string, number]>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     () => ({ foo: 'bar' }),
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<{ foo: string }>(key)
       return () => {}
     }
   )
   useSWRSubscription(
     () => (truthy() ? { foo: 'bar' } : undefined),
-    (key, { next: _ }: SWRSubNext<string, Error>) => {
+    (key, { next: _ }: SWRSubscriptionOptions<string, Error>) => {
       expectType<{ foo: string }>(key)
       return () => {}
     }
